requirements.txt
django==6.0.4
djangorestframework
pytest-django
model-bakery
pyotp
django-redis